src/deployment/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==4.8.0
appdirs==1.4.4
APScheduler==3.9.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
argos-translate-files==1.2.0
argostranslate==1.9.6
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
autocorrect==2.6.1
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.9.3
betterproto==2.0.0b6
bleach==6.2.0
boto3==1.36.20
botocore==1.36.20
cachelib==0.13.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
commonmark==0.9.1
contourpy==1.3.1
contractions==0.1.73
cryptography==44.0.1
ctranslate2==4.5.0
cycler==0.12.1
Cython==3.0.12
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
discord==2.3.2
discord.py==2.4.0
dnspython==2.7.0
docker-pycreds==0.4.0
docopt==0.6.2
dotenv==0.9.9
ekphrasis==0.5.4
email_validator==2.2.0
emoji==2.14.1
executing==2.2.0
expiringdict==1.2.2
fastjsonschema==2.21.1
filelock==3.17.0
Flask==2.2.5
flask-babel==3.1.0
Flask-Limiter==2.6.3
Flask-Session==0.4.0
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.2.0
ftfy==6.3.1
gast==0.4.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googletrans-py==4.0.0
grpcio==1.70.0
grpclib==0.4.7
h11==0.14.0
h2==4.2.0
h5py==3.12.1
hpack==4.1.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.16.4
hyperframe==6.1.0
id==1.5.0
idna==3.10
imageio==2.37.0
in-toto-attestation==0.9.3
invoke==2.2.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kagglehub==0.3.7
keras==3.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
langdetect==1.0.9
lazy_loader==0.4
LexiLang==1.0.4
libclang==18.1.1
limits==4.0.1
llvmlite==0.44.0
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.1.1
ml-dtypes==0.4.1
model-signing==0.2.0
Morfessor==2.0.6
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nlpaug==1.1.11
nltk==3.8.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12>=12.1.3.1
nvidia-cuda-cupti-cu12>=12.1.105
nvidia-cuda-nvrtc-cu12>=12.1.105
nvidia-cuda-runtime-cu12>=12.1.105
nvidia-cudnn-cu12>=8.9.2.26
nvidia-cufft-cu12>=11.0.2.54
nvidia-curand-cu12>=10.3.2.106
nvidia-cusolver-cu12>=11.4.5.107
nvidia-cusparse-cu12>=12.1.0.106
nvidia-cusparselt-cu12>=0.7.1
nvidia-nccl-cu12>=2.19.3
nvidia-nvjitlink-cu12>=12.8.61
nvidia-nvtx-cu12>=12.1.105
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.14.0
overrides==7.7.0
packaging==23.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
polib==1.1.1
prometheus-client==0.15.0
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==3.20.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyahocorasick==2.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
qtconsole==5.6.1
QtPy==2.4.3
redis==4.4.4
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
rfc3161-client==0.1.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc8785==0.1.4
rich==12.6.0
rpds-py==0.22.3
rsa==4.9
s3transfer==0.11.2
sacremoses==0.0.53
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
securesystemslib==1.2.0
Send2Trash==1.8.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
sentry-sdk==2.21.0
setproctitle==1.3.4
shap==0.46.0
sigstore==1.0.0
sigstore-protobuf-specs==0.3.2
sigstore-rekor-types==0.0.18
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
stanza==1.1.1
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
textsearch==0.0.24
tf_keras==2.18.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tinycss2==1.4.0
tokenizers==0.13.3
torch==2.2.2
torch-geometric==2.6.1
torchvision==0.17.2
tornado==6.4.2
tqdm==4.65.2
traitlets==5.14.3
transformers==4.29.0
translatehtml==1.5.2

tuf==5.1.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
ujson==5.10.0
uri-template==1.3.0
urllib3==2.3.0
waitress==2.1.2
wandb==0.19.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.3.8
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.17.2
yarg==0.1.9
yarl==1.18.3
